ai-sales-api/requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
openai==1.11.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
langchain-openai==0.0.5
langchain==0.1.5
